Alter number of threads for xmalloc-test
xmalloc-test takes a parameter for the number of worker threads, `-w N`,
there are two types of threads `allocators` and `releasers`, and it will
allocator the parameters worth of each, i.e. `N`:

    https://github.com/daanx/mimalloc-bench/blob/master/bench/xmalloc-test/xmalloc-test.c#L176-L190

So you expect for an optimal configuration to use

  -w {processors/2}

as the workers parameter for this benchmark.  The test is currently
configured to use

  -w {processors * 2}

workers, and thus (processors * 4) threads.

This commit changes it to uses the same number of threads as processors.