Add snmalloc secure mode
This adds snmalloc's secure mode as a comparison option.  It adds to the
set of secure allocators.

`ssn` didn't read well, so the option is `sn-sec`.

For consistency, we renames the secure versions of mimalloc to have a
`-sec` suffix, rather than a `s` prefix. That is, `smi` has become
`mi-sec`.