Flush required after puts.
`puts` only guarantees to flush it is going to a console like pipe.
The benchmark is piping it to `grep`, and hence it is not flushing.
This meant that snmalloc was getting an artificially high result,
as it was detecting the corruption on teardown, but due to the stream
not being flushed, the `puts` was swallowed by the teardown crash.

Although, I like snmalloc to do well, this doesn't seem in the spirit
of the benchmark.