Add support for PartitionAlloc
This change adds support for PartitionAlloc, the
memory allocator used in chromium. Normally the
allocator is build in-tree, but this pulls in a
standalone version. The standalone version added
here should be updated as soon as the official
repository of PartitionAlloc
(https://chromium.googlesource.com/chromium/src/base/allocator/partition_allocator)
supports standalone builds.